Crate anoma_encoding_spec

Encoding spec markdown generator.

When ran, this generator will:

Get schema for all the types to be included in encoding docs

Merge their definitions

For each schema’s declaration, look-up top-level definitions and format to md table

For each non-top-level definition, format to md table

Rebuild on changes with: cargo watch -x "run --bin anoma_encoding_spec" -i docs

Structs

BORSH_TYPES 🔒

Borsh types may be used by declarations. These are displayed differently in the md_fmt_type.

Table 🔒

Constants

OUTPUT_PATH 🔒

This generator will write output into this docs file.

Functions

definition_to_table 🔒

escape_fragment_anchor 🔒

Escape a link to another type on the page

escape_html 🔒

Escape a type for markdown (rendered as HTML)

main 🔒

md_fmt_type 🔒

Format a type to markdown. For internal types, adds anchors.

write_generated_code_notice 🔒